add to cart and cart page

+++ lib/features/Admin/screens/add_product_screen.dart
@@ -53,7 +53,7 @@ class _AddProductScreenState extends State<AddProductScreen> {
         context: context,
         name: productNameController.text,
         description: descriptionController.text,
-        price: double.parse(priceController.text),
+        price: int.parse(priceController.text),
         quantity: int.parse(quantityController.text),
         category: category,
         images: images,
